% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{economic_growth}
\alias{economic_growth}
\alias{economic_growth_original}
\alias{economic_growth_sala_i_martin}
\title{Economic Growth data}
\format{An object of class \code{data.frame} with 87 rows and 68 columns.}
\usage{
economic_growth

economic_growth_original

economic_growth_sala_i_martin
}
\description{
Sala i Martin's (88 countries)
Leamer's (87 countries)
Original (139 countries)
}
\keyword{datasets}

