% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sValues.R
\name{print.sValues}
\alias{print.sValues}
\title{Succinct display of S-values results.
 
Succinct display of S-values results.}
\usage{
\method{print}{sValues}(x, ..., print.length = 6)
}
\arguments{
\item{x}{an object of class \code{\link{sValues}}.}

\item{...}{further arguments passed to or from other methods.}

\item{print.length}{how many variables to show in the screen?
This is used for pretty printing. The default is 6.

 @return 
 \code{NULL}}
}
\description{
Succinct display of S-values results.
 
Succinct display of S-values results.
}
\examples{
data(economic_growth)
eg_sv <- sValues(GR6096 ~ ., data = economic_growth)
eg_sv
str(eg_sv) 
 
}

