% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sValues.R
\name{str.sValues}
\alias{str.sValues}
\title{str sValues}
\usage{
\method{str}{sValues}(object, max.level = 1, ...)
}
\arguments{
\item{object}{an object of class \code{\link{sValues}}.}

\item{max.level}{maximal level of nesting which is 
applied for displaying nested structures. Default is 1.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{str} method for \code{sValues}.
}

