% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sValues.R
\name{summary.sValues}
\alias{summary.sValues}
\title{summary sValues}
\usage{
\method{summary}{sValues}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{sValues}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
For now, this function is equivalent to \code{\link{print.sValues}}.
}
