% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package_documentation.R
\docType{package}
\name{sValues-package}
\alias{sValues-package}
\title{sValues: measures of the sturdiness of regression coefficients}
\description{
The R package sValues
implements the measure of sturdiness of coefficients proposed by
Leamer (2014) and discussed in Leamer (2015).
The S-values try to provide a sensible framework to assess the sensitivity
of coefficient estimates to model ambiguity.
}
\details{
The main function of the package is the \code{\link{sValues}} function.

More information can be found on its help documentation, examples and vignette.

The package also includes an example dataset on economic growth.
}
\references{
Leamer, Ed. (2014a). S-values: Conventional context-minimal measures of the sturdiness of regression coefficients. Working Paper

Leamer, Ed. (2015). S-values and bayesian weighted all-subsets regressions. European Economic Review.
}

