% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{economic_growth}
\alias{economic_growth}
\alias{economic_growth_original}
\alias{economic_growth_sala_i_martin}
\title{Economic Growth data}
\format{\preformatted{'data.frame':	87 obs. of  68 variables:
 $ GR6096   : num  0.011 0.0011 0.0474 0.0046 0.0024 -0.0252 0.0151 0.0245 0.003 0.0277 ...
 $ ABSLATIT : num  36.72 6.36 21.54 3.37 10.73 ...
 $ AIRDIST  : num  1675 5040 8760 6600 5470 ...
 $ AVELF    : num  0.2937 0.6831 0.3775 0.0133 0.852 ...
 $ BRIT     : num  0 0 1 0 0 0 0 1 0 0 ...
 $ BUDDHA   : num  0 0 0 0 0 0 0 0 0 0 ...
 $ CATH00   : num  0.119 0.01 0.001 0 0.001 0 0.025 0.072 0.368 0.054 ...
 $ CIV72    : num  0.167 0.333 0.5 0 0.5 ...
 $ COLONY   : num  1 1 1 1 1 1 1 1 0 1 ...
 $ CONFUC   : num  0 0 0 0 0 0 0 0 0 0 ...
 $ DENS60   : num  220.53 49.45 1178.23 8.71 87.86 ...
 $ DENS65C  : num  74.7 99.3 0 0 19.6 ...
 $ DENS65I  : num  1.553 12.513 0.969 125.117 11.879 ...
 $ DPOP6090 : num  0.0284 0.024 0.0331 0.0203 0.0263 ...
 $ EAST     : num  0 0 0 0 0 0 0 0 0 0 ...
 $ ECORG    : num  0 0 5 4 5 3 1 1 0 5 ...
 $ ENGFRAC  : num  0 0 0 0 0 0 0 0 0 0 ...
 $ EUROPE   : num  0 0 0 0 0 0 0 0 0 0 ...
 $ FERTLDC1 : num  2 1.92 1.93 1.86 1.74 ...
 $ GDE1     : num  0.03 0.018 0.035 0.014 0.024 ...
 $ GDPCH60L : num  7.45 7 6.28 6.46 6.46 ...
 $ GEEREC1  : num  0.0297 0.0248 0.03 0.0183 0.0221 ...
 $ GGCFD3   : num  0.23 0.02 0.08 0.03 0.04 0.06 0.04 0.01 0.01 0.13 ...
 $ GOVNOM1  : num  0.15 0.11 0.21 0.04 0.16 0.21 0.18 0.21 0.1 0.15 ...
 $ GOVSH61  : num  0.17 0.28 0.2 0.04 0.19 0.36 0.32 0.28 0.16 0.09 ...
 $ GVR61    : num  0.107 0.237 0.134 0.01 0.144 ...
 $ H60      : num  0 0 0 0 0 0 0 0.05 0 0 ...
 $ HERF00   : num  0.764 0.848 0.755 0.996 0.899 ...
 $ HINDU00  : num  0 0 0 0 0 0 0 0 0 0 ...
 $ IPRICE1  : num  98 150 63 115 140 ...
 $ LAAM     : num  0 0 0 0 0 0 0 0 0 0 ...
 $ LANDAREA : num  2381740 110620 566730 25680 465400 ...
 $ LANDLOCK : num  0 0 1 1 0 1 0 0 1 0 ...
 $ LHCPC    : num  6.51 1.05 -4.61 -4.61 3.89 ...
 $ LIFE060  : num  47.3 38.9 45.7 41.8 43.4 39.3 47.3 46.4 42.2 40.9 ...
 $ LT100CR  : num  0.05 0.11 0 0 0.13 0 0.05 0.39 0.02 0.27 ...
 $ MALFAL66 : num  0.09 1 0.76 1 1 1 1 0.028 0.85 1 ...
 $ MINING   : num  0.2 0.01 0.53 0.01 0.09 0.03 0.17 0.03 0 0.22 ...
 $ MUSLIM00 : num  0.866 0.07 0 0.002 0.05 0.004 0 0.811 0.26 0 ...
 $ NEWSTATE : num  2 2 2 2 2 2 2 1 0 2 ...
 $ OIL      : num  1 0 0 0 0 0 0 0 0 1 ...
 $ OPENDEC1 : num  0.469 0.441 0.852 0.222 0.434 ...
 $ ORTH00   : num  0 0 0 0 0 0 0 0.066 0.363 0 ...
 $ OTHFRAC  : num  0.836 0 0 0 0.007 ...
 $ P60      : num  0.46 0.27 0.42 0.18 0.65 0.32 0.78 0.66 0.07 1 ...
 $ PI6090   : num  6.45 4.4 7.56 7.3 6.89 ...
 $ SQPI6090 : num  41.6 19.3 57.2 53.3 47.5 ...
 $ PRIGHTS  : num  5.83 7 2.06 6.83 6.06 6.83 6.56 5.11 6.67 6 ...
 $ POP1560  : num  0.437 0.407 0.476 0.405 0.405 ...
 $ POP60    : num  10800 2237 481 2948 5297 ...
 $ POP6560  : num  0.0385 0.0729 0.0333 0.0302 0.0347 ...
 $ PRIEXP70 : num  0.933 0.896 0.75 0.977 0.915 0.557 0.71 0.729 0.983 0.91 ...
 $ PROT00   : num  0.003 0.002 0.142 0 0.002 0 0 0.114 0 0.011 ...
 $ RERD     : num  190 126 131 167 187 156 191 168 98 169 ...
 $ REVCOUP  : num  0.12 0.1 0 0.39 0.08 0.39 0.36 0.16 0.72 0.12 ...
 $ SAFRICA  : num  0 1 1 1 1 1 1 0 1 1 ...
 $ SCOUT    : num  1 0 0 0 1 0 1 1 0 0 ...
 $ SIZE60   : num  16.7 14.7 12.5 14.5 15 ...
 $ SOCIALIST: num  1 1 0 0 0 0 1 0 1 0 ...
 $ SPAIN    : num  0 0 0 0 0 0 0 0 0 0 ...
 $ TOT1DEC1 : num  0.1001 -0.0244 -0.0492 -0.0692 0.014 ...
 $ TOTIND   : num  0.3586 0.1241 0.0138 0.1586 0.3241 ...
 $ TROPICAR : num  0.162 1 0.703 1 1 ...
 $ TROPPOP  : num  0.129 0.514 0.5 1 0.864 ...
 $ WARTIME  : num  0 0 0 0.04 0 0 0 0.08 1 0 ...
 $ WARTORN  : num  1 0 0 1 0 0 0 1 1 0 ...
 $ YRSOPEN  : num  0 0.11 0.36 0 0.16 0 0 0 0 0 ...
 $ ZTROPICS : num  0 0.97971 0.00837 0 0.24094 ...
}}
\usage{
economic_growth

economic_growth_original

economic_growth_sala_i_martin
}
\description{
Sala i Martin's (88 countries)
Leamer's (87 countries)
Original (139 countries)
}
\keyword{datasets}

